Add files via upload
Hey, so i just created this branch because i dont wanna to change anything in the master until its fixed and works. I made a couple of changes and attempted to change all the strings to QStrings. Also, instead of having multiple windows, and create more files than we need, i figure out how to work the stacked widget, so that way we only have one mainWindow to work with. Somehow the program is giving an error, when casting the QString to int, we have store in one of our dates, from sales.cpp. Maybe you guys can see whats going on.